Refactor adapter.
Move into util package. Remove circular dependencies, add nullability annotations, fix Javadoc.
