+++ spring-data-mongodb/src/main/java/org/springframework/data/mongodb/util/MongoClientVersion.java
@@ -36,7 +36,9 @@ public class MongoClientVersion {
 	private static final boolean REACTIVE_CLIENT_PRESENT = ClassUtils
 			.isPresent("com.mongodb.reactivestreams.client.MongoClient", MongoClientVersion.class.getClassLoader());
 
-	private static final boolean IS_5PlusClient = ClassUtils.isPresent("com.mongodb.connection.StreamFactoryFactory", MongoClientVersion.class.getClassLoader());
+	private static final boolean IS_VERSION_5_OR_NEWER = ClassUtils
+			.isPresent("com.mongodb.internal.connection.StreamFactoryFactory", MongoClientVersion.class.getClassLoader());
+
 	/**
 	 * @return {@literal true} if the async MongoDB Java driver is on classpath.
 	 */
@@ -60,7 +62,11 @@ public class MongoClientVersion {
 		return REACTIVE_CLIENT_PRESENT;
 	}
 
-	public static boolean is5PlusClient() {
-		return IS_5PlusClient;
+	/**
+	 * @return {@literal true} if the MongoDB Java driver version is 5 or newer.
+	 * @since 4.3
+	 */
+	public static boolean isVersion5OrNewer() {
+		return IS_VERSION_5_OR_NEWER;
 	}
 }
